sdks/official/fraiseql-python/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "pytest>=9.0.2",
  "anyio[trio]>=4.0",
  "pytest-anyio>=0.0.0",
  "httpx>=0.27",
  "ruff>=0.8.0",
  "ty>=0.0.1a8"
]

[project]
authors = [
  {name = "FraiseQL Team"}
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Database :: Front-Ends"
]
dependencies = [
  "httpx>=0.27"
]
description = "FraiseQL v2 - Compiled GraphQL execution engine (schema authoring)"
keywords = ["graphql", "database", "sql", "compiler", "schema"]
license = {text = "MIT"}
name = "fraiseql"
readme = "README.md"
requires-python = ">=3.10"
version = "2.0.1.dev0"

[project.optional-dependencies]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]

[project.urls]
Documentation = "https://fraiseql.readthedocs.io"
Homepage = "https://github.com/fraiseql/fraiseql"
Issues = "https://github.com/fraiseql/fraiseql/issues"
Repository = "https://github.com/fraiseql/fraiseql"

[tool.hatch.build.targets.wheel]
packages = ["src/fraiseql"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
ignore = [
  "ANN401"  # typing.Any usage allowed for dynamic types, **kwargs, runtime introspection
]
# Select rule categories:
# E - pycodestyle errors
# F - pyflakes
# I - isort (import sorting)
# N - pep8-naming
# W - pycodestyle warnings
# UP - pyupgrade (modern Python syntax)
# ANN - flake8-annotations (type annotations)
# B - flake8-bugbear (likely bugs)
# C4 - flake8-comprehensions (better comprehensions)
# SIM - flake8-simplify (simplification suggestions)
# TCH - flake8-type-checking (type checking imports)
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "C4", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"src/fraiseql/integrations/langchain.py" = [
  "TC002"  # Runtime try/except imports cannot be moved to TYPE_CHECKING
]
"src/fraiseql/integrations/llamaindex.py" = [
  "TC001"  # Runtime import used in __init__ constructor
]
"tests/**/*.py" = [
  "ANN"  # Don't require type annotations in tests
]
